backend/requirements.txt
fastapi
uvicorn
groq
sentence-transformers
faiss-cpu
python-dotenv


